changed cleanup to skip currently-indexed files in xml

+++ mycleanup.py
@@ -4,11 +4,13 @@ import re
 from mydeploy import (
     connect_to_bucket,
     get_aws_credentials,
+    get_file_objects,
     )
 
 # all paths are relative to jenkins job's workspace (absolute path is fine too)
 AWS_CONFIG_PATH = ''    # path to config file (ini format) containing aws credentials
 AWS_PROFILE = ''        # name of aws profile from the config file to be used
+XML_PATH = ''           # path of the xml file containing latest file versions
 
 # name of buckets for connection purposes
 CSS_BUCKET = ''
@@ -30,15 +32,24 @@ def cleanup_main():
     js_bucket = connect_to_bucket(cred, JS_BUCKET)
     image_bucket = connect_to_bucket(cred, IMAGE_BUCKET)
 
-    css = [css_bucket, CSS_PREFIX]
-    js = [js_bucket, JS_PREFIX]
-    image = [image_bucket, IMAGE_PREFIX]
+    existing_versioned_files_in_xml = get_file_objects(AWS_CONFIG_PATH, AWS_PROFILE,
+                                                       CSS_BUCKET, JS_BUCKET, IMAGE_BUCKET,
+                                                       XML_PATH)
 
-    for bucket in [css, js, image]:
-        keys_to_delete = get_all_matching_keys(bucket[0], bucket[1])
-        for key_ in keys_to_delete:
-            key_.delete()
-            print('Deleted http://' + bucket[0].name + '.s3.amazonaws.com/' + key_.key + '\n')
+    keys_in_xml = [item.versioned_path_in_bucket for item in existing_versioned_files_in_xml]
+
+    for bucket in [[css_bucket, CSS_PREFIX], [js_bucket, JS_PREFIX], [image_bucket, IMAGE_PREFIX]]:
+        keys_matching_pattern = get_all_matching_keys(bucket[0], bucket[1])
+
+        for key_ in keys_matching_pattern:
+
+            if (key_.key in keys_in_xml):
+                print('Skipping deletion of http://' + bucket[0].name + '.s3.amazonaws.com/' + key_.key +
+                      ', currently indexed in XML file \n')
+
+            else:
+                key_.delete()
+                print('Deleted http://' + bucket[0].name + '.s3.amazonaws.com/' + key_.key + '\n')
 
 
 def get_all_matching_keys(bucket, prefix_=None):
